Escape JS-forbidden unicode chars to be nice to broken JS libs.
Javascript does not allow U+2028 and U+2029 in strings, whereas JSON
does. This is the one place where JSON is not a strict subset of JS,
but is not widely known, so many JS libs incorrectly use eval to
decode JSON. Always encoding these characters allows compatibility with
these (broken) libraries.